# file-uploading-app/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://0684-190-64-82-98.ngrok-free.app'
paths:
  /api/csrf-cookie:
    get:
      summary: 'Return an empty response simply to trigger the storage of the CSRF cookie in the browser.'
      operationId: returnAnEmptyResponseSimplyToTriggerTheStorageOfTheCSRFCookieInTheBrowser
      description: ''
      parameters: []
      responses:
        204:
          description: ''
      tags:
        - Endpoints
      security: []
  '/api/files/{fileId}/versions':
    post:
      summary: 'Add New Version'
      operationId: addNewVersion
      description: 'Add a new version to an existing file.'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'New version added successfully'
                  version:
                    id: 1
                    version_number: 2
                    path: files/1/version_2.pdf
                    download_url: 'http://example.com/api/files/versions/1/download'
                    created_at: '2024-01-27T12:00:00Z'
                    updated_at: '2024-01-27T12:00:00Z'
                properties:
                  message:
                    type: string
                    example: 'New version added successfully'
                  version:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      version_number:
                        type: integer
                        example: 2
                      path:
                        type: string
                        example: files/1/version_2.pdf
                      download_url:
                        type: string
                        example: 'http://example.com/api/files/versions/1/download'
                      created_at:
                        type: string
                        example: '2024-01-27T12:00:00Z'
                      updated_at:
                        type: string
                        example: '2024-01-27T12:00:00Z'
      tags:
        - 'File Management'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 'The new version file. Must not exceed 10MB.'
                  nullable: false
              required:
                - file
      security: []
    get:
      summary: 'Get File Versions'
      operationId: getFileVersions
      description: 'Retrieve all versions of a specific file.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      version_number: 2
                      path: files/1/version_2.pdf
                      download_url: 'http://example.com/api/files/versions/1/download'
                      comments_count: 3
                      created_at: '2024-01-27T12:00:00Z'
                      updated_at: '2024-01-27T12:00:00Z'
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        version_number: 2
                        path: files/1/version_2.pdf
                        download_url: 'http://example.com/api/files/versions/1/download'
                        comments_count: 3
                        created_at: '2024-01-27T12:00:00Z'
                        updated_at: '2024-01-27T12:00:00Z'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        version_number:
                          type: integer
                          example: 2
                        path:
                          type: string
                          example: files/1/version_2.pdf
                        download_url:
                          type: string
                          example: 'http://example.com/api/files/versions/1/download'
                        comments_count:
                          type: integer
                          example: 3
                        created_at:
                          type: string
                          example: '2024-01-27T12:00:00Z'
                        updated_at:
                          type: string
                          example: '2024-01-27T12:00:00Z'
      tags:
        - 'File Management'
      security: []
    parameters:
      -
        in: path
        name: fileId
        description: 'The ID of the file.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/files/versions/{versionId}/comments':
    post:
      summary: 'Add Version Comment'
      operationId: addVersionComment
      description: 'Add a comment to a specific file version.'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Comment added successfully'
                  comment:
                    id: 1
                    comment: 'This version includes the requested changes.'
                    created_at: '2024-01-27T12:00:00Z'
                    updated_at: '2024-01-27T12:00:00Z'
                properties:
                  message:
                    type: string
                    example: 'Comment added successfully'
                  comment:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      comment:
                        type: string
                        example: 'This version includes the requested changes.'
                      created_at:
                        type: string
                        example: '2024-01-27T12:00:00Z'
                      updated_at:
                        type: string
                        example: '2024-01-27T12:00:00Z'
      tags:
        - 'File Management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  description: 'The comment text. Must not exceed 1000 characters.'
                  example: 'This version includes the requested changes.'
                  nullable: false
              required:
                - comment
      security: []
    parameters:
      -
        in: path
        name: versionId
        description: 'The ID of the version.'
        example: 1
        required: true
        schema:
          type: integer
  '/api/files/versions/{versionId}/download':
    get:
      summary: 'Download Version File'
      operationId: downloadVersionFile
      description: 'Download a specific version of a file.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: file
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'You do not have permission to download this file'
                properties:
                  message:
                    type: string
                    example: 'You do not have permission to download this file'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Version not found'
                properties:
                  message:
                    type: string
                    example: 'Version not found'
      tags:
        - 'File Management'
      security: []
    parameters:
      -
        in: path
        name: versionId
        description: 'The ID of the version to download.'
        example: 1
        required: true
        schema:
          type: integer
  /api/files/latest:
    get:
      summary: 'Get Latest Uploads'
      operationId: getLatestUploads
      description: 'Retrieve the latest file uploads with pagination.'
      parameters:
        -
          in: query
          name: limit
          description: 'The number of items per page. Default: 10'
          example: 15
          required: false
          schema:
            type: integer
            description: 'The number of items per page. Default: 10'
            example: 15
            nullable: false
        -
          in: query
          name: page
          description: 'The page number. Default: 1'
          example: 1
          required: false
          schema:
            type: integer
            description: 'The page number. Default: 1'
            example: 1
            nullable: false
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      name: document.pdf
                      extension: pdf
                      size_formatted: '1.00 MB'
                      created_at: '2024-01-27T12:00:00Z'
                  meta:
                    current_page: 1
                    last_page: 5
                    per_page: 10
                    total: 50
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        name: document.pdf
                        extension: pdf
                        size_formatted: '1.00 MB'
                        created_at: '2024-01-27T12:00:00Z'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: document.pdf
                        extension:
                          type: string
                          example: pdf
                        size_formatted:
                          type: string
                          example: '1.00 MB'
                        created_at:
                          type: string
                          example: '2024-01-27T12:00:00Z'
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 5
                      per_page:
                        type: integer
                        example: 10
                      total:
                        type: integer
                        example: 50
      tags:
        - 'File Management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  description: 'Must be at least 1. Must not be greater than 50.'
                  example: 9
                  nullable: false
                page:
                  type: integer
                  description: 'Must be at least 1.'
                  example: 20
                  nullable: false
      security: []
tags:
  -
    name: Endpoints
    description: ''
  -
    name: 'File Management'
    description: "\nAPIs for managing files, versions and comments"